+++ app/controllers/hosted_events/checkin_controller.rb
@@ -1,6 +1,6 @@
 class HostedEvents::CheckinController < ApplicationController
   include SetsEvent
-
+  include MarkPaid
 
   before_action :load_attendance, only: [:update]
 
@@ -29,25 +29,7 @@ class HostedEvents::CheckinController < ApplicationController
   end
 
 
-  private
 
-
-  def mark_paid
-    payment_method = params[:payment_method]
-    if Payable::Methods::ALL.include?(payment_method)
-
-      @attendance.mark_orders_as_paid!(check_number: params[:check_number])
-
-      if @attendance.owes_nothing?
-        return render file: '/hosted_events/checkin/mark_paid'
-      else
-        return render file: '/hosted_events/checkin/error.js.erb'
-      end
-    end
-
-    # if something didn't flow correctly, just don't do anything
-    render nothing: true
-  end
 
   def checkin_and_crossover
     @attendance.crossover_orientation = params[:competition][:orientation]
